+++ lib/src/platform/linux/endian.c
@@ -82,7 +82,7 @@ uint16_t cio_htole16(uint16_t host_endian_16bits)
 	return htole16(host_endian_16bits);
 }
 
-uint32_t cio_htole32(uint16_t host_endian_32bits)
+uint32_t cio_htole32(uint32_t host_endian_32bits)
 {
 	return htole32(host_endian_32bits);
 }
